+++ CHANGELOG.md
@@ -5,6 +5,7 @@ https://semver.org/
 * Trace outline performance improvements
 * MOCA Connection bug fix
 * SQL Formatter bug fix -- https://github.com/mrglassdanny/moca-language-server/issues/63
+* SQL/Groovy compilation bug fix -- https://github.com/mrglassdanny/moca-language-server/issues/61
 
 ## 1.7.14 (December 15, 2020)
 * Trace outline tweaks